MR-2

+++ src/db/ClientManagerEventFlag.cpp
@@ -51,7 +51,7 @@ void CClientManager::SetEventFlag(TPacketSetEventFlag* p)
 		char szQuery[1024];
 		snprintf(szQuery, sizeof(szQuery),
 				"REPLACE INTO quest%s (dwPID, szName, szState, lValue) VALUES(0, '%s', '', %ld)",
-				GetTablePostfix(), p->szFlagName, p->lValue);
+				GetTablePostfix(), p->szFlagName, static_cast<long>(p->lValue));
 		szQuery[1023] = '\0';
 
 		//CDBManager::instance().ReturnQuery(szQuery, QID_QUEST_SAVE, 0, NULL);
