+++ src/game/MarkManager.cpp
@@ -18,7 +18,7 @@ void CGuildMarkManager::__DeleteImage(CGuildMarkImage * pkImgDel)
 CGuildMarkManager::CGuildMarkManager()
 {
 	// 남은 mark id 셋을 만든다. (서버용)
-	for (DWORD i = 0; i < MAX_IMAGE_COUNT * CGuildMarkImage::MARK_TOTAL_COUNT; ++i)
+	for (DWORD i = 0; i < (DWORD)MAX_IMAGE_COUNT * (DWORD)CGuildMarkImage::MARK_TOTAL_COUNT; ++i)
 		m_setFreeMarkID.insert(i);
 }
 
@@ -103,7 +103,7 @@ void CGuildMarkManager::LoadMarkImages()
 	{
 		DWORD markID = it->second;
 
-		if (markID < MAX_IMAGE_COUNT * CGuildMarkImage::MARK_TOTAL_COUNT)
+		if (markID < (DWORD)MAX_IMAGE_COUNT * (DWORD)CGuildMarkImage::MARK_TOTAL_COUNT)
 			isMarkExists[markID / CGuildMarkImage::MARK_TOTAL_COUNT] = true;
 	}
 
@@ -151,7 +151,7 @@ CGuildMarkImage * CGuildMarkManager::__GetImage(DWORD imgIdx)
 
 bool CGuildMarkManager::AddMarkIDByGuildID(DWORD guildID, DWORD markID)
 {
-	if (markID >= MAX_IMAGE_COUNT * CGuildMarkImage::MARK_TOTAL_COUNT)
+	if (markID >= (DWORD)MAX_IMAGE_COUNT * (DWORD)CGuildMarkImage::MARK_TOTAL_COUNT)
 		return false;
 
 	//sys_log(0, "MarkManager: guild_id=%d mark_id=%d", guildID, markID);
